src/week-2-august-8th-august-14th/matrix-rank-transform.ts
/*
    Given an m x n matrix, return a new matrix answer where answer[row][col] is the rank of matrix[row][col].

    The rank is an integer that represents how large an element is compared to other elements. It is calculated using the following rules:

    The rank is an integer starting from 1.
    If two elements p and q are in the same row or column, then:
    If p < q then rank(p) < rank(q)
    If p == q then rank(p) == rank(q)
    If p > q then rank(p) > rank(q)
    The rank should be as small as possible.
    It is guaranteed that answer is unique under the given rules.

    Constraints:

    m == matrix.length
    n == matrix[i].length
    1 <= m, n <= 500
    -109 <= matrix[row][col] <= 109
 */

export function matrixRankTransform(matrix: number[][]): number[][] {
    // Sort the cells by value and process them in increasing order.

    // The rank of a cell is the maximum rank in its row and column plus one.

    // Handle the equal cells by treating them as components using a union-find data structure.

    return [
        [1, 2],
        [2, 3]
    ];
}
